requirements.txt
pycryptodome
matplotlib
numpy